+++ server/db/user/user.go
@@ -1,4 +1,4 @@
-package auth
+package user
 
 import (
 	"database/sql"
